addressed comments and restore template
