+++ nautobot/apps/views.py
@@ -22,7 +22,6 @@ from nautobot.core.views.mixins import (
     NautobotViewSetMixin,
     ObjectBulkCreateViewMixin,  # 3.0 TODO: deprecated, will be removed in 3.0
     ObjectBulkDestroyViewMixin,
-    ObjectBulkDisconnectViewMixin,
     ObjectBulkUpdateViewMixin,
     ObjectChangeLogViewMixin,
     ObjectDataComplianceViewMixin,
@@ -72,7 +71,6 @@ __all__ = (
     "ObjectApprovalWorkflowView",
     "ObjectBulkCreateViewMixin",  # 3.0 TODO: remove this
     "ObjectBulkDestroyViewMixin",
-    "ObjectBulkDisconnectViewMixin",
     "ObjectBulkUpdateViewMixin",
     "ObjectChangeLogViewMixin",
     "ObjectDataComplianceViewMixin",
