+++ nautobot/dcim/templates/dcim/consoleport_retrieve.html
@@ -47,10 +47,12 @@
                     <td>Cable</td>
                     <td>
                         {{ object.cable|hyperlinked_object }}
-                        <a href="{% url 'dcim:consoleport_trace' pk=object.pk %}" class="btn btn-primary btn-sm" title="Trace">
-                            <span class="mdi mdi-transit-connection-variant" aria-hidden="true"></span>
-                        </a>
+                        {% include 'dcim/inc/cable_actions.html' with trace_url_name='dcim:consoleport_trace' %}
                     </td>
+                </tr>
+                <tr>
+                    <td>Cable Status</td>
+                    <td>{{ object.cable.status|hyperlinked_object_with_color }}</td>
                 </tr>
                 {% if object.connected_endpoint %}
                     <tr>
@@ -90,17 +92,7 @@
             <div class="card-body text-secondary">
                 Not connected
                 {% if perms.dcim.add_cable %}
-                    <span class="dropdown float-end">
-                        {# comment The "fixed" strategy allows the dropdown to break out of the containing card #}
-                        <button type="button" class="btn btn-primary btn-sm dropdown-toggle" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false" data-bs-popper-config='{"strategy": "fixed"}'>
-                            <span class="mdi mdi-ethernet-cable" aria-hidden="true"></span> Connect
-                        </button>
-                        <ul class="dropdown-menu dropdown-menu-end">
-                            <li><a class="dropdown-item" href="{% url 'dcim:consoleport_connect' termination_a_id=object.pk termination_b_type='console-server-port' %}?return_url={{ object.get_absolute_url }}">Console Server Port</a></li>
-                            <li><a class="dropdown-item" href="{% url 'dcim:consoleport_connect' termination_a_id=object.pk termination_b_type='front-port' %}?return_url={{ object.get_absolute_url }}">Front Port</a></li>
-                            <li><a class="dropdown-item" href="{% url 'dcim:consoleport_connect' termination_a_id=object.pk termination_b_type='rear-port' %}?return_url={{ object.get_absolute_url }}">Rear Port</a></li>
-                        </ul>
-                    </span>
+                    {% include 'dcim/inc/cable_add_button.html' with termination=object %}
                 {% endif %}
             </div>
         {% endif %}
